Money format, change § to ChatColor

+++ node/src/main/kotlin/me/slavita/construction/worker/Worker.kt
@@ -11,6 +11,6 @@ class Worker(
 ) {
     val uuid = UUID.randomUUID()
 
-    val sellPrice: Int
+    val sellPrice: Long
         get() { return 199 /* Формула */ }
 }
